feat: messages

+++ pineapple-ai/src/main/java/com/cboy/pineapple/ai/utils/AssistantMessageDiagnostic.java
@@ -5,4 +5,4 @@ import java.util.Optional;
 
 record DiagnosticErrorInfo(Optional<String> name, String message, Optional<String> stack, Optional<String> code) {}
 
-public record AssistantMessageDiagnostic(String type, double timestamp, Optional<DiagnosticErrorInfo> error, Map<String, Object> details) {}
+public record AssistantMessageDiagnostic(String type, long timestamp, Optional<DiagnosticErrorInfo> error, Map<String, Object> details) {}
